+++ .gitignore
@@ -5,3 +5,6 @@ __pycache__/
 
 venv/
 myenv/
+
+.vscode/
+*.code-workspace
