+++ README.md
@@ -1,5 +1,5 @@
-# Pyutils - Simplifying Python Development
-"Pyutils" is a versatile utility package that provides a set of modules designed to streamline various aspects of Python development. Whether you're working on mathematical computations, managing input/output, handling files, or organizing data collections, "Pyutils" has you covered.
+# PyDevTools - Simplifying Python Development
+"PyDevTools" is a versatile utility package that provides a set of modules designed to streamline various aspects of Python development. Whether you're working on mathematical computations, managing input/output, handling files, or organizing data collections, "PyDevTools" has you covered.
 
 ### Features
 - Math: Mathematical operations made easy with a range of functions and utilities.
@@ -10,31 +10,31 @@
 - Database: Interact with databases seamlessly.
 
 ### Installation
-To use "Pyutils" in your project, you can install it via pip:
+To use "PyDevTools" in your project, you can install it via pip:
 
 ```bash
-pip install git+https://github.com/JeelDobariya38/Pyutils.git
+pip install git+https://github.com/JeelDobariya38/PyDevTools.git
 ```
 
 ### Uninstalling
-To uninstall "Pyutils," use the following command:
+To uninstall "PyDevTools," use the following command:
 ```bash
-pip uninstall Pyutils
+pip uninstall PyDevTools
 ```
 
 ### Updating
-To update "Pyutils" to the latest version, run:
+To update "PyDevTools" to the latest version, run:
 
 ```bash
-pip install --upgrade git+https://github.com/JeelDobariya38/Pyutils.git
+pip install --upgrade git+https://github.com/JeelDobariya38/PyDevTools.git
 ```
 
 ### Usage
-Here's a quick example of how to use the "Pyutils" package:
+Here's a quick example of how to use the "PyDevTools" package:
 
 ```python
-# Import the math module from Pyutils
-from Pyutils.Math import math
+# Import the math module from PyDevTools
+from PyDevTools.Math import math
 
 # Use a function from the math module
 result = math.add(5, 3)
@@ -43,7 +43,7 @@ print(result)  # Output: 8
 For detailed information on each module and its functions, please refer to the documentation.
 
 ### Contributing
-Contributions to "Pyutils" are welcome! If you have ideas for improvements, bug fixes, or new features, please feel free to submit a pull request or open a new issue in repository.
+Contributions to "PyDevTools" are welcome! If you have ideas for improvements, bug fixes, or new features, please feel free to submit a pull request or open a new issue in repository.
 
 ### License
 This project is licensed under the MIT License.
