+++ Tests/pydevtools/math/test_math.py
@@ -1,5 +1,5 @@
 import pytest
-from Pyutils.Math import math
+from pydevtools.math import math
 
 # Testing the add function
 def test_add_integers():
