+++ docs/index.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Home | Pyutils</title>
+    <title>Home | PyDevTools</title>
     <style>
         ul li {
             list-style: circle;
@@ -13,13 +13,13 @@
 </head>
 
 <body>
-    <h1>Pyutils - Simplifying Python Development</h1>
+    <h1>PyDevTools - Simplifying Python Development</h1>
     <p>
-        "Pyutils" is a versatile utility package that provides a set of modules designed to streamline various aspects
+        "PyDevTools" is a versatile utility package that provides a set of modules designed to streamline various aspects
         of Python development.
         Whether you're working on mathematical computations, managing input/output, handling files, or organizing data
         collections,
-        "Pyutils" has you covered.
+        "PyDevTools" has you covered.
     </p>
     <p>
         <h2>Features</h2>
@@ -34,11 +34,11 @@
     </p>
     <p>
         <h2>Installation</h2>
-        To use "Pyutils" in your project, you can install it via pip:
+        To use "PyDevTools" in your project, you can install it via pip:
         
         <div style="background: #000; color: rgb(0, 252, 168); padding: 5px;">
         ```bash <br>
-        pip install git+https://github.com/JeelDobariya38/Pyutils.git <br>
+        pip install git+https://github.com/JeelDobariya38/PyDevTools.git <br>
         ```
         </div>
     </p>
